requirements.txt
# Web Framework
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-multipart==0.0.9

# Database
sqlalchemy[asyncio]==2.0.35
asyncpg==0.29.0
alembic==1.13.2

# Cache
redis==5.1.1

# Data Validation
pydantic==2.9.2
pydantic-settings==2.5.2

# AI Engine
dashscope==1.20.11
httpx==0.27.2

# Template & Config
jinja2==3.1.4
pyyaml==6.0.2

# Frontend
streamlit==1.38.0
plotly==5.24.1

# Utilities
python-dotenv==1.0.1
python-jose[cryptography]==3.3.0

# Testing
pytest==8.3.3
pytest-asyncio==0.24.0
